Moved data methods back to AllenNlpTestCase
